5.20 수정 1. 회원 탈퇴시 진행이 되지 않는 오류 수정 2. 팔로우 알림 수정

+++ backend/server/module/firebaseProcess.js
@@ -28,13 +28,20 @@ var checkID = async (token) => {
 var deleteFeed = async (token) => {
   var query = db.collection('feed').where('UserToken', '==', token);
   query.get().then((snapshot) => {
-    snapshot.forEach((doc) => {
-      doc.ref.delete();
-    });
+    if (snapshot != undefined) {
+      snapshot.forEach((doc) => {
+        doc.ref.delete();
+      });
+    }
   });
 };
 
-var showlist = async (token) => {
+// var deleteList = async(token) =>{
+//   var query = db.collection("users").document(token).
+// }
+
+//다른 사용자의 팔로워/팔로잉 리스트에서 전달된 사용자의 데이터를 제거한다. 
+var cleanFollowlist = async (token) => {
   var followerquery = db.collectionGroup('follower').where('token', '==', token);
   var followingquery = db.collectionGroup('following').where('token', '==', token);
   followerquery.get().then((snapshot) => {
@@ -62,9 +69,20 @@ var showlist = async (token) => {
         });
     });
   });
-  return true;
 }
 
-exports.checkID = checkID;
-exports.deleteFeed = deleteFeed;
-exports.showlist = showlist;
+//사용자가 참가한 챌린지
+
+
+//사용자의 데이터를 지우는 과정 
+
+var cleanUserData = async (token) => {
+  checkID(token).then((data) => {
+    if (data) cleanFollowlist(token).then(() => {
+      deleteFeed(token);
+    });
+  });
+}
+
+
+exports.cleanUserData = cleanUserData;
